Moved Pillar2Error from "controllers" to "models"
