+++ app/uk/gov/hmrc/pillar2submissionapi/services/ObligationsAndSubmissionsService.scala
@@ -21,7 +21,7 @@ import play.api.libs.json.Format.GenericFormat
 import play.api.libs.json.{JsError, JsSuccess}
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import uk.gov.hmrc.pillar2submissionapi.connectors.ObligationAndSubmissionsConnector
-import uk.gov.hmrc.pillar2submissionapi.controllers.error.{DownstreamValidationError, UnexpectedResponse}
+import uk.gov.hmrc.pillar2submissionapi.models.error.Pillar2Error.{DownstreamValidationError, UnexpectedResponse}
 import uk.gov.hmrc.pillar2submissionapi.models.obligationsandsubmissions.{ObligationsAndSubmissionsErrorResponse, ObligationsAndSubmissionsSuccessResponse}
 
 import java.time.LocalDate
