+++ app/uk/gov/hmrc/pillar2submissionapi/services/OverseasReturnNotificationService.scala
@@ -21,7 +21,7 @@ import play.api.Logging
 import play.api.libs.json.{JsError, JsSuccess, Json}
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import uk.gov.hmrc.pillar2submissionapi.connectors.OverseasReturnNotificationConnector
-import uk.gov.hmrc.pillar2submissionapi.controllers.error.{DownstreamValidationError, ORNNotFoundException, UnexpectedResponse}
+import uk.gov.hmrc.pillar2submissionapi.models.error.Pillar2Error.{DownstreamValidationError, ORNNotFoundException, UnexpectedResponse}
 import uk.gov.hmrc.pillar2submissionapi.models.overseasreturnnotification._
 
 import scala.concurrent.{ExecutionContext, Future}
