+++ test/uk/gov/hmrc/pillar2submissionapi/controllers/AccountActivityControllerSpec.scala
@@ -29,9 +29,10 @@ import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.pillar2submissionapi.base.ControllerBaseSpec
 import uk.gov.hmrc.pillar2submissionapi.config.AppConfig
 import uk.gov.hmrc.pillar2submissionapi.controllers.accountactivity.AccountActivityController
-import uk.gov.hmrc.pillar2submissionapi.controllers.error._
 import uk.gov.hmrc.pillar2submissionapi.helpers.AccountActivityDataFixture
 import uk.gov.hmrc.pillar2submissionapi.models.accountactivity.AccountActivitySuccessResponse
+import uk.gov.hmrc.pillar2submissionapi.models.error.Pillar2Error
+import uk.gov.hmrc.pillar2submissionapi.models.error.Pillar2Error._
 import uk.gov.hmrc.pillar2submissionapi.services.AccountActivityService
 
 import scala.concurrent.Future
