+++ test/uk/gov/hmrc/pillar2submissionapi/services/SubmitBTNServiceSpec.scala
@@ -23,9 +23,9 @@ import play.api.libs.json.Json
 import play.api.test.Helpers.{await, defaultAwaitTimeout}
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import uk.gov.hmrc.pillar2submissionapi.base.UnitTestBaseSpec
-import uk.gov.hmrc.pillar2submissionapi.controllers.error._
 import uk.gov.hmrc.pillar2submissionapi.models.belowthresholdnotification.{BTNSubmission, SubmitBTNSuccessResponse}
 import uk.gov.hmrc.pillar2submissionapi.models.btn.{BTNSuccess, BTNSuccessResponse}
+import uk.gov.hmrc.pillar2submissionapi.models.error.Pillar2Error._
 import uk.gov.hmrc.pillar2submissionapi.models.hip.{ApiFailure, ApiFailureResponse}
 import uk.gov.hmrc.pillar2submissionapi.services.SubmitBTNServiceSpec._
